@@ -4,7 +4,7 @@
   "description": "modular SCSS Boilerplate",
   "main": "Gulpfile.babel.js",
   "scripts": {
-    "start": "node_modules/gulp/bin/gulp.js",
+    "start": "node ./node_modules/gulp/bin/gulp.js",
     "test": "npm test"
   },
   "repository": {
@@ -35,17 +35,22 @@
     "babel-preset-env": "^1.7.0",
     "babel-preset-stage-0": "^6.24.1",
     "babel-register": "^6.26.0",
+    "chalk": "^2.4.1",
     "del": "^3.0.0",
-    "eslint": "^1.10.3",
+    "eslint": "^2.13.1",
+    "eslint-config-google": "^0.10.0",
+    "eslint-config-standard": "^12.0.0",
     "eslint-config-standard-babel": "0.0.2",
     "eslint-plugin-babel": "^4.1.2",
     "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx": "0.0.2",
     "eslint-plugin-node": "^7.0.1",
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-react": "^7.11.1",
     "eslint-plugin-standard": "^4.0.0",
     "gulp": "^3.9.1",
     "gulp-notify": "^3.2.0",
+    "gulp-rename": "^1.4.0",
     "gulp-sass": "^4.0.1",
     "node-sass": "^4.9.3",
     "run-sequence": "^2.2.1"
