@@ -1,5 +1,5 @@
-export default (function(config) {
-    const rootRel   = '../../'
+export default (function (config) {
+    const rootRel   = process.cwd() + '/'
 
     const paths = {
         src:    rootRel + 'web/src/',
@@ -9,12 +9,19 @@ export default (function(config) {
     config      = Object.assign({
         paths,
         globs: {
+            clean:  [
+                '!**/.gitkeep', 'stack/__test__/**/*', 'web/dist/**/*'
+            ],
             scss:   {
-                dist:   paths.dist + 'scss/',
-                src:    [
-                    paths.src + 'full.scss',
-                    paths.src + 'mq/*.scss'
-                ]
+                watch:      paths.src + '**/*.scss',
+                src:    {
+                    main:   paths.src + 'full.scss',
+                    mqs:    paths.src + 'mqs/*.scss'
+                },
+                dist:   {
+                    main:   paths.dist + 'css/',
+                    mqs:    paths.dist + 'css/mqs/'
+                }
             }
         }
     }, config) // assign process.env.APP_CONFIG to defaults
