@@ -15,29 +15,23 @@ const taskFiles = fs.readdirSync('./stack/gulp/gulp-tasks/')
         /(\.(js|coffee)$)/i.test(path.extname(name))
     )
 
-
 // SET THE ENVIRONMENT VARIABLES (mainly NODE_ENV)
 // run config before importing/requiring the other tasks
 // because they import appConfig which depends on config-module
 require('./gulp-tasks/set-env.js')
 runSequence('set-env') // = sets NODE_ENV based on --env
 
-
 // IMPORT TASK FILES
 taskFiles.forEach(taskFile => {
     require('./gulp-tasks/' + taskFile)
 })
 
-// TBD: gulp task list
-//
-// require('gulp-task-list')(gulp, null, [__filename, ...tasks.map(filename => 'gulp/tasks/' + filename)])
-// gulp.task('default', cb => runSequence('task-list'))
-
 gulp.task('default', done => {
     const gulpSequence = [
         'set-env',
-        'clean'
-    // '    compile'
+        'clean',
+        'sass:build-main',
+        'sass:watch'
     ]
 
     // if (process.env.NODE_ENV !== 'development') {
